// prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const JobStatus = {
  Available: 'Available',
  Executing: 'Executing',
  Retryable: 'Retryable',
  Completed: 'Completed',
  Discarded: 'Discarded'
} as const

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus]